+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/repository/SysRoleSystemAttributeRepository.java
@@ -4,7 +4,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 import eu.bcvsolutions.idm.acc.dto.filter.RoleSystemAttributeFilter;
 import eu.bcvsolutions.idm.acc.entity.SysRoleSystem;
@@ -18,13 +17,6 @@ import eu.bcvsolutions.idm.core.api.repository.AbstractEntityRepository;
  * @author svandav
  *
  */
-@RepositoryRestResource(//
-		collectionResourceRel = "roleSystemAttributes", //
-		path = "role-system-attributes", //
-		itemResourceRel = "roleSystemAttribute", //
-		exported = false // we are using repository metadata, but we want expose
-							// rest endpoint manually
-)
 public interface SysRoleSystemAttributeRepository
 		extends AbstractEntityRepository<SysRoleSystemAttribute, RoleSystemAttributeFilter> {