+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/repository/SysSyncItemLogRepository.java
@@ -3,7 +3,6 @@ package eu.bcvsolutions.idm.acc.repository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 import eu.bcvsolutions.idm.acc.dto.filter.SyncItemLogFilter;
 import eu.bcvsolutions.idm.acc.entity.SysSyncItemLog;
@@ -15,13 +14,6 @@ import eu.bcvsolutions.idm.core.api.repository.AbstractEntityRepository;
  * @author Svanda
  *
  */
-@RepositoryRestResource(//
-		collectionResourceRel = "syncItemLogs", //
-		path = "sync-item-logs", //
-		itemResourceRel = "syncActionLog", //
-		exported = false // we are using repository metadata, but we want expose
-							// rest endpoint manually
-)
 public interface SysSyncItemLogRepository extends AbstractEntityRepository<SysSyncItemLog, SyncItemLogFilter> {
 	@Override
 	@Query(value = "select e from SysSyncItemLog e"+